test: create stock test mixin for assertion/utils
